msds-434/application/requirements.txt
Flask==0.12.1
google-api-python-client==1.6.2
google-cloud-automl
requests
requests-toolbelt